Split out last loader
