+++ packages/blade/private/shell/loaders.ts
@@ -1,4 +1,4 @@
-import { cp, readFile } from 'node:fs/promises';
+import { cp, readFile, rename } from 'node:fs/promises';
 import path from 'node:path';
 import { compile } from '@mdx-js/mdx';
 import withToc from '@stefanprobst/rehype-extract-toc';
@@ -16,6 +16,7 @@ import {
   exists,
   extractDeclarationName,
   getFileList,
+  prepareStyles,
   wrapClientExport,
 } from '@/private/shell/utils';
 import {
@@ -25,6 +26,7 @@ import {
 } from '@/private/shell/utils/providers';
 import type { DeploymentProvider } from '@/private/universal/types/util';
 import { generateUniqueId } from '@/private/universal/utils/crypto';
+import { getOutputFile } from '@/private/universal/utils/paths';
 
 export const getClientReferenceLoader = (): esbuild.Plugin => ({
   name: 'Client Reference Loader',
@@ -262,6 +264,51 @@ export const getProviderLoader = (
           await transformToVercelBuildOutput();
           break;
         }
+      }
+    });
+  },
+});
+
+export const getMetaLoader = (
+  environment: 'development' | 'production',
+  projects: Array<string>,
+): esbuild.Plugin => ({
+  name: 'Init Loader',
+  setup(build) {
+    let bundleId: string | undefined;
+
+    build.onStart(() => {
+      bundleId = generateUniqueId();
+    });
+
+    build.onResolve({ filter: /^build-meta$/ }, (source) => ({
+      path: source.path,
+      namespace: 'dynamic-meta',
+    }));
+
+    build.onLoad({ filter: /^build-meta$/, namespace: 'dynamic-meta' }, () => ({
+      contents: `export const bundleId = "${bundleId}";`,
+      loader: 'ts',
+      resolveDir: process.cwd(),
+    }));
+
+    build.onEnd(async (result) => {
+      if (result.errors.length === 0) {
+        const clientBundle = path.join(outputDirectory, getOutputFile('init', 'js'));
+        const clientSourcemap = path.join(
+          outputDirectory,
+          getOutputFile('init', 'js.map'),
+        );
+
+        await Promise.all([
+          rename(clientBundle, clientBundle.replace('init.js', `${bundleId}.js`)),
+          rename(
+            clientSourcemap,
+            clientSourcemap.replace('init.js.map', `${bundleId}.js.map`),
+          ),
+        ]);
+
+        await prepareStyles(environment, projects, bundleId as string);
       }
     });
   },
