+++ packages/blade/private/shell/utils/build.ts
@@ -1,4 +1,4 @@
-import { cp, readFile, rename } from 'node:fs/promises';
+import { readFile, rename } from 'node:fs/promises';
 import path from 'node:path';
 import * as esbuild from 'esbuild';
 
@@ -6,13 +6,13 @@ import {
   clientInputFile,
   defaultDeploymentProvider,
   outputDirectory,
-  publicDirectory,
   serverInputFolder,
 } from '@/private/shell/constants';
 import {
   getClientReferenceLoader,
   getFileListLoader,
   getMdxLoader,
+  getMetaLoader,
   getProviderLoader,
   getReactAriaLoader,
 } from '@/private/shell/loaders';
@@ -21,12 +21,7 @@ import {
   exists,
   prepareStyles,
 } from '@/private/shell/utils';
-import {
-  getProvider,
-  transformToCloudflareOutput,
-  transformToNetlifyOutput,
-  transformToVercelBuildOutput,
-} from '@/private/shell/utils/providers';
+import { getProvider } from '@/private/shell/utils/providers';
 import { generateUniqueId } from '@/private/universal/utils/crypto';
 import { getOutputFile } from '@/private/universal/utils/paths';
 
@@ -54,8 +49,6 @@ export const build = async (
       );
     }
   }
-
-  let bundleId: string | undefined;
 
   const entryPoints: esbuild.BuildOptions['entryPoints'] = [
     {
@@ -92,48 +85,7 @@ export const build = async (
       getMdxLoader('production'),
       getReactAriaLoader(),
       getClientReferenceLoader(),
-      {
-        name: 'Init Loader',
-        setup(build) {
-          build.onStart(() => {
-            bundleId = generateUniqueId();
-          });
-
-          build.onResolve({ filter: /^build-meta$/ }, (source) => ({
-            path: source.path,
-            namespace: 'dynamic-meta',
-          }));
-
-          build.onLoad({ filter: /^build-meta$/, namespace: 'dynamic-meta' }, () => ({
-            contents: `export const bundleId = "${bundleId}";`,
-            loader: 'ts',
-            resolveDir: process.cwd(),
-          }));
-
-          build.onEnd(async (result) => {
-            if (result.errors.length === 0) {
-              const clientBundle = path.join(
-                outputDirectory,
-                getOutputFile('init', 'js'),
-              );
-              const clientSourcemap = path.join(
-                outputDirectory,
-                getOutputFile('init', 'js.map'),
-              );
-
-              await Promise.all([
-                rename(clientBundle, clientBundle.replace('init.js', `${bundleId}.js`)),
-                rename(
-                  clientSourcemap,
-                  clientSourcemap.replace('init.js.map', `${bundleId}.js.map`),
-                ),
-              ]);
-
-              await prepareStyles(environment, projects, bundleId as string);
-            }
-          });
-        },
-      },
+      getMetaLoader(environment, projects),
       getProviderLoader(environment, provider),
       ...(options?.plugins || []),
     ],
